pyproject.toml
[tool.poetry]
name = "home-agent-suite"
version = "0.1.0"
description = "A multi-agent, multi-tool bot for home automation."
authors = [ "David Snider <david@davidsnider.org>" ]
readme = "README.md"
package-mode = false

# Add a new dependencies section for shared dev tools
[tool.poetry.dependencies]
python = "^3.11"
tzlocal = "^5.3.1"
pulumi = "^3.187.0"
pulumi-github = "^6.7.3"
streamlit = "^1.48.0"
google-adk = "^1.11.0"
python-dotenv = "^1.1.1"
geopy = "^2.4.1"
tomorrow-io-client = {path = "libs/tomorrow_io_client", develop = true}
# Agent dependencies - only supervisor as it's a proper package
supervisor-agent = {path = "agents/supervisor", develop = true}
# Shared libraries
common-logging = {path = "libs/common_logging", develop = true}
home-assistant-client = {path = "libs/home_assistant_client", develop = true}

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
black = "^26.3.1"
flake8 = "^7.0.0"
detect-secrets = "^1.5.0"
pytest-cov = "^7.0.0"
requests-mock = "^1.12.1"
pytest-asyncio = "^1.2.0"
aioresponses = "^0.7.4"
pytest-mock = "^3.10.0"
pytest-env = "^1.2.0"
respx = "^0.22.0"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
exclude = ".venv,**/.venv,**/__pycache__/*,**/*.pyc"
